Backend/requirements.txt
flask
httpx
python-dotenv
tldextract
pymysql
cryptography